+++ common/artifact-manager/src/test/java/com/vmware/pscoe/iac/artifact/store/vrang/VraNgBlueprintStoreTest.java
@@ -22,7 +22,6 @@ import com.vmware.pscoe.iac.artifact.configuration.ConfigurationVraNg;
 import com.vmware.pscoe.iac.artifact.helpers.AssertionsHelper;
 import com.vmware.pscoe.iac.artifact.helpers.FsMocks;
 import com.vmware.pscoe.iac.artifact.helpers.stubs.BlueprintMockBuilder;
-import com.vmware.pscoe.iac.artifact.helpers.stubs.BlueprintVersionsMockBuilder;
 import com.vmware.pscoe.iac.artifact.model.Package;
 import com.vmware.pscoe.iac.artifact.model.PackageFactory;
 import com.vmware.pscoe.iac.artifact.model.PackageType;
@@ -49,8 +48,6 @@ import static org.mockito.Mockito.*;
 
 /**
  * NOTE: This does not test duplicate names from one content source, since the Store is not responsible for that kind of logic.
- *
- * @TODO ADD assertions for versions when exporting
  */
 public class VraNgBlueprintStoreTest {
 
@@ -125,9 +122,7 @@ public class VraNgBlueprintStoreTest {
 		store.exportContent();
 
 		String[] expectedBlueprints = {"ngnix"};
-		String[] expectedBlueprintFiles = {"content.yaml",
-			"details.json",
-			"versions.json"};
+		String[] expectedBlueprintFiles = {"content.yaml","details.json"};
 
 		// VERIFY
 		AssertionsHelper.assertFolderContainsFiles(fsMocks.getTempFolderProjectPath(), expectedBlueprints);
@@ -154,9 +149,7 @@ public class VraNgBlueprintStoreTest {
 		store.exportContent();
 
 		String[] expectedBlueprints = {"ngnix"};
-		String[] expectedBlueprintFiles = {"content.yaml",
-			"details.json",
-			"versions.json"};
+		String[] expectedBlueprintFiles = {"content.yaml","details.json"};
 
 		// VERIFY
 		AssertionsHelper.assertFolderContainsFiles(fsMocks.getTempFolderProjectPath(), expectedBlueprints);
@@ -211,53 +204,7 @@ public class VraNgBlueprintStoreTest {
 	}
 
 	@Test
-	void testImportContentFlipsVersions() throws IOException, ParseException {
-		// GIVEN
-		List<String> blueprintNames = new ArrayList<>();
-		blueprintNames.add("nginx");
-
-		when(vraNgPackageDescriptor.getBlueprint()).thenReturn(blueprintNames);
-
-		BlueprintMockBuilder builder = new BlueprintMockBuilder("nginx");
-		VraNgBlueprint blueprint = builder.build();
-		BlueprintVersionsMockBuilder versionsMockBuilder = new BlueprintVersionsMockBuilder(blueprint);
-		versionsMockBuilder.setVersions(3);
-		JsonArray versionsArray = versionsMockBuilder.build();
-
-		fsMocks.blueprintFsMocks().addBlueprint(blueprint, versionsArray.toString());
-
-		AssertionsHelper.assertFolderContainsFiles(fsMocks.getTempFolderProjectPath(), new String[]{"nginx"});
-		AssertionsHelper.assertFolderContainsFiles(fsMocks.findItemByNameInFolder(fsMocks.getTempFolderProjectPath(), "nginx"), new String[]{"content.yaml", "details.json", "versions.json"});
-
-		List<VraNgBlueprint> bluePrintsOnServer = new ArrayList<>();
-
-		when(restClient.getAllBlueprints()).thenReturn(bluePrintsOnServer);
-
-		// START TEST
-		store.importContent(tempFolder.getRoot());
-
-		// VERIFY
-		verify(restClient, times(1)).createBlueprint(any());
-
-		ArgumentCaptor<String> bpIdCaptor = ArgumentCaptor.forClass(String.class);
-		ArgumentCaptor<JsonObject> jsonObjectCaptor = ArgumentCaptor.forClass(JsonObject.class);
-
-		verify(restClient, times(3)).importBlueprintVersion(
-			bpIdCaptor.capture(),
-			jsonObjectCaptor.capture()
-		);
-
-		Date first = this.getDateFromJsonElement(jsonObjectCaptor.getAllValues().get(0).get("createdAt"));
-		Date second = this.getDateFromJsonElement(jsonObjectCaptor.getAllValues().get(1).get("createdAt"));
-		Date third = this.getDateFromJsonElement(jsonObjectCaptor.getAllValues().get(2).get("createdAt"));
-
-		// Assert ordering
-		assertEquals(-1, first.compareTo(second));
-		assertEquals(-1, second.compareTo(third));
-	}
-
-	@Test
-	void testImportContentWithNoVersions() throws IOException, ParseException {
+	void testImportContent() throws IOException, ParseException {
 		// GIVEN
 		List<String> blueprintNames = new ArrayList<>();
 		blueprintNames.add("nginx");
@@ -286,79 +233,6 @@ public class VraNgBlueprintStoreTest {
 	}
 
 	@Test
-	void testImportContentFlipsVersionsWithDescending() throws IOException, ParseException {
-		// GIVEN
-		List<String> blueprintNames = new ArrayList<>();
-		blueprintNames.add("nginx");
-
-		when(vraNgPackageDescriptor.getBlueprint()).thenReturn(blueprintNames);
-
-		BlueprintMockBuilder builder = new BlueprintMockBuilder("nginx");
-		VraNgBlueprint blueprint = builder.build();
-		BlueprintVersionsMockBuilder versionsMockBuilder = new BlueprintVersionsMockBuilder(blueprint);
-		versionsMockBuilder.setVersions(3);
-		JsonArray tempArray = versionsMockBuilder.build();
-		JsonArray versionsArray = new JsonArray();
-		versionsArray.add(tempArray.get(2));
-		versionsArray.add(tempArray.get(1));
-		versionsArray.add(tempArray.get(0));
-
-		fsMocks.blueprintFsMocks().addBlueprint(blueprint, versionsArray.toString());
-
-		AssertionsHelper.assertFolderContainsFiles(fsMocks.getTempFolderProjectPath(), new String[]{"nginx"});
-		AssertionsHelper.assertFolderContainsFiles(fsMocks.findItemByNameInFolder(fsMocks.getTempFolderProjectPath(), "nginx"), new String[]{"content.yaml", "details.json", "versions.json"});
-
-		List<VraNgBlueprint> bluePrintsOnServer = new ArrayList<>();
-
-		when(restClient.getAllBlueprints()).thenReturn(bluePrintsOnServer);
-
-		// START TEST
-		store.importContent(tempFolder.getRoot());
-
-		// VERIFY
-		verify(restClient, times(1)).createBlueprint(any());
-
-		ArgumentCaptor<String> bpIdCaptor = ArgumentCaptor.forClass(String.class);
-		ArgumentCaptor<JsonObject> jsonObjectCaptor = ArgumentCaptor.forClass(JsonObject.class);
-
-		verify(restClient, times(3)).importBlueprintVersion(
-			bpIdCaptor.capture(),
-			jsonObjectCaptor.capture()
-		);
-
-		// Assert ordering
-		Date first = this.getDateFromJsonElement(jsonObjectCaptor.getAllValues().get(0).get("createdAt"));
-		Date second = this.getDateFromJsonElement(jsonObjectCaptor.getAllValues().get(1).get("createdAt"));
-		Date third = this.getDateFromJsonElement(jsonObjectCaptor.getAllValues().get(2).get("createdAt"));
-
-		// Assert ordering
-		assertEquals(-1, first.compareTo(second));
-		assertEquals(-1, second.compareTo(third));
-	}
-
-	@Test
-	void testExportContentWithoutVersions() {
-		// GIVEN
-		List<VraNgBlueprint> blueprints = new ArrayList<>();
-		BlueprintMockBuilder mockBuilder = new BlueprintMockBuilder("ngnix");
-
-		blueprints.add(mockBuilder.setRequestScopeOrg(true).build());
-
-		when(vraNgPackageDescriptor.getBlueprint()).thenReturn(null);
-		when(restClient.getAllBlueprints()).thenReturn(blueprints);
-		when(config.getIgnoreBlueprintVersions()).thenReturn(true);
-
-		// TEST
-		store.exportContent();
-
-		String[] expectedBlueprintFiles = {"content.yaml", "details.json"};
-
-		// VERIFY
-		AssertionsHelper.assertFolderContainsFiles(
-			fsMocks.findItemByNameInFolder(fsMocks.getTempFolderProjectPath(), "ngnix"), expectedBlueprintFiles);
-	}
-
-	@Test
 	void testImportContentForNonExistingBlueprintsInConfiguration() throws IOException, ParseException {
 		// GIVEN
 		List<String> blueprintNames = new ArrayList<>();
@@ -368,13 +242,11 @@ public class VraNgBlueprintStoreTest {
 
 		BlueprintMockBuilder builder = new BlueprintMockBuilder("nginx");
 		VraNgBlueprint blueprint = builder.build();
-		BlueprintVersionsMockBuilder versionsMockBuilder = new BlueprintVersionsMockBuilder(blueprint);
-		versionsMockBuilder.setVersions(3);
-		JsonArray versionsArray = versionsMockBuilder.build();
-		fsMocks.blueprintFsMocks().addBlueprint(blueprint, versionsArray.toString());
 
+		fsMocks.blueprintFsMocks().addBlueprint(blueprint);
+		
 		AssertionsHelper.assertFolderContainsFiles(fsMocks.getTempFolderProjectPath(), new String[]{"nginx"});
-		AssertionsHelper.assertFolderContainsFiles(fsMocks.findItemByNameInFolder(fsMocks.getTempFolderProjectPath(), "nginx"), new String[]{"content.yaml", "details.json", "versions.json"});
+		AssertionsHelper.assertFolderContainsFiles(fsMocks.findItemByNameInFolder(fsMocks.getTempFolderProjectPath(), "nginx"), new String[]{"content.yaml", "details.json"});
 
 		List<VraNgBlueprint> bluePrintsOnServer = new ArrayList<>();
 
@@ -391,25 +263,5 @@ public class VraNgBlueprintStoreTest {
 
 		// VERIFY
 		verify(restClient, times(1)).createBlueprint(any());
-		ArgumentCaptor<String> bpIdCaptor = ArgumentCaptor.forClass(String.class);
-		ArgumentCaptor<JsonObject> jsonObjectCaptor = ArgumentCaptor.forClass(JsonObject.class);
-
-		verify(restClient, times(3)).importBlueprintVersion(
-			bpIdCaptor.capture(),
-			jsonObjectCaptor.capture()
-		);
-
-		Date first = this.getDateFromJsonElement(jsonObjectCaptor.getAllValues().get(0).get("createdAt"));
-		Date second = this.getDateFromJsonElement(jsonObjectCaptor.getAllValues().get(1).get("createdAt"));
-		Date third = this.getDateFromJsonElement(jsonObjectCaptor.getAllValues().get(2).get("createdAt"));
-
-		// Assert ordering
-		assertEquals(-1, first.compareTo(second));
-		assertEquals(-1, second.compareTo(third));
-	}
-
-	private Date getDateFromJsonElement(JsonElement jsonElement) throws ParseException {
-		return new SimpleDateFormat(BlueprintVersionsMockBuilder.DATE_FORMAT).parse(
-			jsonElement.getAsJsonPrimitive().toString().replaceAll("^\"|\"$", ""));
 	}
 }